backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
psycopg2-binary==2.9.9
djoser==2.2.0
djangorestframework_simplejwt==5.3.0
django-colorfield==0.10.1
django-filter==23.3
drf_extra_fields==3.7.0
pillow==10.1.0
python-dotenv==1.0.0
sorl-thumbnail==12.10.0